Added the "activation" of the celsius link to the "search" function
Added this so that, when the user clicks the search button, the celsius link is activated (this means that it will have the dark blue color). This is an important detail because the user might convert the temperature to F for a certain city and then search another city, and the API response gives the temperature for that last city in celsius. Without what I added in this commit, the F link would still be active.